+++ src/packages/auth/__init__.py
@@ -1,2 +1,3 @@
-from .utils import get_current_user, get_password_hash
+from .utils import get_current_user, get_password_hash, create_access_token
 from .router import auth_router
+from .responses import Token